+++ core-ng/src/test/java/core/framework/impl/web/HTTPServerHandlerTest.java
@@ -15,7 +15,7 @@ class HTTPServerHandlerTest {
 
     @BeforeEach
     void createHTTPServerHandler() {
-        handler = new HTTPServerHandler(null, null, null, null);
+        handler = new HTTPServerHandler(null, null, null);
     }
 
     @Test